Consolidates evm based reencryption conditions to eventually encapsulate payment methods; Reduce scope of conditions to a module
