+++ nucypher/cli/types.py
@@ -29,7 +29,7 @@ from nucypher.blockchain.economics import Economics
 from nucypher.blockchain.eth.interfaces import BlockchainInterface
 from nucypher.blockchain.eth.networks import NetworksInventory
 from nucypher.blockchain.eth.token import TToken
-from nucypher.policy.conditions import PAYMENT_METHODS
+from nucypher.policy.payment import PAYMENT_METHODS
 from nucypher.utilities.networking import validate_operator_ip, InvalidOperatorIP
